fix: update error-collector imports
